+++ index.html
@@ -10,8 +10,7 @@
 </head>
 
 <body>
-<script src="/js/functions.js"></script>
-<script src="/js/main.js"></script>
+
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
@@ -235,6 +234,9 @@
         <h2 class="data-error__title">Не удалось загрузить данные</h2>
       </section>
     </template>
+
+  <script src="./js/main.js" type="module"></script>
+  <script src="./js/functions.js"></script>
 
 </body>
 </html>
