+++ js/data.js
@@ -1,0 +1,49 @@
+
+const DESCRIPTIONS = [
+  'Селфи (фотография самого себя)',
+  'Пейзаж (природа, город)',
+  'Животное (изображение питомца или дикого зверя)',
+  'Предмет (фото любимой вещи)',
+  'Мем (смешная картинка)',
+  'Коллаж (сочетание нескольких элементов)',
+  'Арт (рисованное изображение)'
+];
+
+const NAMES = [
+  'Олег',
+  'Хлоя',
+  'Лео',
+  'Изабелла',
+  'Оливер',
+  'София',
+  'Джеймс',
+  'Миа',
+  'Ноа',
+  'Эмили',
+  'Бен',
+  'Лина',
+  'Маркус',
+  'Сара',
+  'Антон',
+  'Ханна',
+  'Крис',
+  'Лея',
+  'Адам',
+  'Елена',
+  'Эрик',
+  'Нина',
+  'Павел',
+  'Феликс',
+  'Джина'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export { DESCRIPTIONS, NAMES, MESSAGES };
